# docs/api/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is an API for the GoChat",
        "title": "GoChat API",
        "contact": {},
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:3000",
    "basePath": "/api/v1",
    "paths": {
...
        "/guild": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Guild"
                ],
                "summary": "Create guild",
                "parameters": [
                    {
                        "description": "Guild data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/cmd_api_endpoints_guild.UpdateGuildRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Guild",
                        "schema": {
                            "$ref": "#/definitions/github_com_FlameInTheDark_gochat_internal_dto.Guild"
                        }
                    },
                    "400": {
                        "description": "Incorrect request body",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Something bad happened",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/guild/{guild_id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Guild"
                ],
                "summary": "Get guild",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Guild id",
                        "name": "guild_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Guild",
                        "schema": {
                            "$ref": "#/definitions/github_com_FlameInTheDark_gochat_internal_dto.Guild"
                        }
                    },
                    "400": {
                        "description": "Incorrect request body",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Something bad happened",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "patch": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Guild"
                ],
                "summary": "Update guild",
                "parameters": [
                    {
                        "description": "Update guild data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/cmd_api_endpoints_guild.UpdateGuildRequest"
                        }
                    },
                    {
                        "type": "integer",
                        "description": "Guild ID",
                        "name": "guild_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Guild",
                        "schema": {
                            "$ref": "#/definitions/github_com_FlameInTheDark_gochat_internal_dto.Guild"
                        }
                    },
                    "400": {
                        "description": "Incorrect request body",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Something bad happened",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/guild/{guild_id}/category": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Guild"
                ],
                "summary": "Create guild channel category",
                "parameters": [
                    {
                        "description": "Create category data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/cmd_api_endpoints_guild.CreateGuildChannelCategoryRequest"
                        }
                    },
                    {
                        "type": "integer",
                        "description": "Guild ID",
                        "name": "guild_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Incorrect request body",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Something bad happened",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/guild/{guild_id}/category/{category_id}": {
            "delete": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Guild"
                ],
                "summary": "Delete channel category",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Guild ID",
                        "name": "guild_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Category ID (actually a channel with special type)",
                        "name": "category_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Deleted",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Incorrect request body",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Something bad happened",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/guild/{guild_id}/channel": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Guild"
                ],
                "summary": "Get guild channels",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Guild id",
                        "name": "guild_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of channels",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_FlameInTheDark_gochat_internal_dto.Channel"
                            }
                        }
                    },
                    "400": {
                        "description": "Incorrect request body",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Something bad happened",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Guild"
                ],
                "summary": "Add guild channel category",
                "parameters": [
                    {
                        "description": "Create category data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/cmd_api_endpoints_guild.CreateGuildChannelCategoryRequest"
                        }
                    },
                    {
                        "type": "integer",
                        "description": "Guild ID",
                        "name": "guild_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Incorrect request body",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Something bad happened",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/guild/{guild_id}/channel/{channel_id}": {
            "delete": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Guild"
                ],
                "summary": "Delete channel",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Guild ID",
                        "name": "guild_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Channel ID",
                        "name": "channel_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Deleted",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Incorrect request body",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Something bad happened",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "cmd_api_endpoints_guild.CreateGuildChannelCategoryRequest": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "private": {
                    "type": "boolean"
                }
            }
        },
        "cmd_api_endpoints_guild.UpdateGuildRequest": {
            "type": "object",
            "properties": {
                "icon_id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "public": {
                    "type": "boolean"
                }
            }
        },
...
        "github_com_FlameInTheDark_gochat_internal_database_model.ChannelType": {
            "type": "integer",
            "enum": [
                0,
                1,
                2,
                3,
                4,
                5
            ],
            "x-enum-varnames": [
                "ChannelTypeGuild",
                "ChannelTypeGuildVoice",
                "ChannelTypeGuildCategory",
                "ChannelTypeDM",
                "ChannelTypeGroupDM",
                "ChannelTypeThread"
            ]
        },
...
        "github_com_FlameInTheDark_gochat_internal_dto.Channel": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "guild_id": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "parent_id": {
                    "type": "integer"
                },
                "permissions": {
                    "type": "integer"
                },
                "position": {
                    "type": "integer"
                },
                "topic": {
                    "type": "string"
                },
                "type": {
                    "$ref": "#/definitions/github_com_FlameInTheDark_gochat_internal_database_model.ChannelType"
                }
            }
        },
        "github_com_FlameInTheDark_gochat_internal_dto.Guild": {
            "type": "object",
            "properties": {
                "icon": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "owner": {
                    "type": "integer"
                },
                "public": {
                    "type": "boolean"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "basic"
        }
    }
}